Fixed IL2CPP crash by removing `AsyncLock` and `AsyncMonitor` on Unity versions older than 2021.2. Implemented `IEquatable<Key>` on `AsyncLock.Key`.
