+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/AsyncLockInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63a9a88b2a007bc4b882b09978f289b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
