+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Threading/AsyncMonitor.cs
@@ -3,9 +3,15 @@
 #endif
 
 using System.Diagnostics;
+using System.Runtime.CompilerServices;
 
 namespace Proto.Promises.Threading
 {
+
+#if UNITY_2021_2_OR_NEWER || NETSTANDARD2_1_OR_GREATER || NETCOREAPP
+    // Old IL2CPP runtime crashes if this code exists, even if it is not used. So we only include them in newer build targets that old Unity versions cannot consume.
+    // See https://github.com/timcassell/ProtoPromise/issues/169 for details.
+
     /// <summary>
     /// Async-compatible Monitor.
     /// </summary>
@@ -20,6 +26,7 @@ namespace Proto.Promises.Threading
         /// </summary>
         /// <param name="asyncLock">The async lock instance that is being entered.</param>
         /// <param name="cancelationToken">The <see cref="CancelationToken"/> used to cancel the lock. If the token is canceled before the lock has been acquired, the returned <see cref="Promise{T}"/> will be canceled.</param>
+        [MethodImpl(Internal.InlineOption)]
         public static Promise<AsyncLock.Key> EnterAsync(AsyncLock asyncLock, CancelationToken cancelationToken = default(CancelationToken))
         {
             return asyncLock.LockAsync(cancelationToken);
@@ -30,6 +37,7 @@ namespace Proto.Promises.Threading
         /// </summary>
         /// <param name="asyncLock">The async lock instance that is being entered.</param>
         /// <param name="cancelationToken">The <see cref="CancelationToken"/> used to cancel the lock. If the token is canceled before the lock has been acquired, <see cref="CanceledException"/> will be thrown.</param>
+        [MethodImpl(Internal.InlineOption)]
         public static AsyncLock.Key Enter(AsyncLock asyncLock, CancelationToken cancelationToken = default(CancelationToken))
         {
             return asyncLock.Lock(cancelationToken);
@@ -41,6 +49,7 @@ namespace Proto.Promises.Threading
         /// <param name="asyncLock">The async lock instance that is being entered.</param>
         /// <param name="key">If successful, the key that will release the lock when it is disposed.</param>
         /// <returns>True if the lock was acquired, false otherwise.</returns>
+        [MethodImpl(Internal.InlineOption)]
         public static bool TryEnter(AsyncLock asyncLock, out AsyncLock.Key key)
         {
             return asyncLock.TryEnter(out key);
@@ -49,6 +58,7 @@ namespace Proto.Promises.Threading
         /// <summary>
         /// Release the lock on the <see cref="AsyncLock"/> associated with the <paramref name="asyncLockKey"/>.
         /// </summary>
+        [MethodImpl(Internal.InlineOption)]
         public static void Exit(AsyncLock.Key asyncLockKey)
         {
             asyncLockKey.Dispose();
@@ -65,6 +75,7 @@ namespace Proto.Promises.Threading
         /// Its result will be true if the lock was re-acquired before the <paramref name="cancelationToken"/> was canceled,
         /// false if the lock was re-acquired after the <paramref name="cancelationToken"/> was canceled
         /// </returns>
+        [MethodImpl(Internal.InlineOption)]
         public static Promise<bool> WaitAsync(AsyncLock.Key asyncLockKey, CancelationToken cancelationToken = default(CancelationToken))
         {
             return asyncLockKey.WaitAsync(cancelationToken);
@@ -80,6 +91,7 @@ namespace Proto.Promises.Threading
         /// true if the lock was re-acquired before the <paramref name="cancelationToken"/> was canceled,
         /// false if the lock was re-acquired after the <paramref name="cancelationToken"/> was canceled
         /// </returns>
+        [MethodImpl(Internal.InlineOption)]
         public static bool Wait(AsyncLock.Key asyncLockKey, CancelationToken cancelationToken = default(CancelationToken))
         {
             return asyncLockKey.Wait(cancelationToken);
@@ -89,6 +101,7 @@ namespace Proto.Promises.Threading
         /// Send a signal to a single waiter on the <see cref="AsyncLock"/> associated with the <paramref name="asyncLockKey"/>.
         /// </summary>
         /// <param name="asyncLockKey">The key to the <see cref="AsyncLock"/> that is currently acquired.</param>
+        [MethodImpl(Internal.InlineOption)]
         public static void Pulse(AsyncLock.Key asyncLockKey)
         {
             asyncLockKey.Pulse();
@@ -98,9 +111,12 @@ namespace Proto.Promises.Threading
         /// Send a signal to all waiters on the <see cref="AsyncLock"/> associated with the <paramref name="asyncLockKey"/>.
         /// </summary>
         /// <param name="asyncLockKey">The key to the <see cref="AsyncLock"/> that is currently acquired.</param>
+        [MethodImpl(Internal.InlineOption)]
         public static void PulseAll(AsyncLock.Key asyncLockKey)
         {
             asyncLockKey.PulseAll();
         }
     } // class AsyncMonitor
-} // namespace Proto.Promises
+
+#endif // UNITY_2021_2_OR_NEWER || NETSTANDARD2_1_OR_GREATER || NETCOREAPP
+} // namespace Proto.Promises.Threading
