+++ ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/UnityTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d21b805f6725a5c45bbc47fbf676cba7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
